GradeBook
The Gradebook takes values from the console, calculates the highest grade, lowest grade and the average grade and then using pattern matching return an average for the Gradebook in the console window. A basic application but helps me to understand the basics of C# and OOP within the CLR environment.